+++ src/crimson/os/seastore/journal.cc
@@ -50,8 +50,8 @@ segment_nonce_t generate_nonce(
     sizeof(meta.seastore_id.uuid));
 }
 
-Journal::Journal(SegmentManager &segment_manager)
-  : segment_manager(segment_manager) {}
+Journal::Journal(SegmentManager &segment_manager, Scanner& scanner)
+  : segment_manager(segment_manager), scanner(scanner) {}
 
 
 Journal::initialize_segment_ertr::future<segment_seq_t>
@@ -97,35 +97,6 @@ Journal::initialize_segment(Segment &segment)
     crimson::ct_error::assert_all{
       "Invalid error in Journal::initialize_segment"
     });
-}
-
-bool Journal::validate_metadata(const bufferlist &bl)
-{
-  auto bliter = bl.cbegin();
-  auto test_crc = bliter.crc32c(
-    ceph::encoded_sizeof_bounded<record_header_t>(),
-    -1);
-  ceph_le32 recorded_crc_le;
-  decode(recorded_crc_le, bliter);
-  uint32_t recorded_crc = recorded_crc_le;
-  test_crc = bliter.crc32c(
-    bliter.get_remaining(),
-    test_crc);
-  return test_crc == recorded_crc;
-}
-
-Journal::read_validate_data_ret Journal::read_validate_data(
-  paddr_t record_base,
-  const record_header_t &header)
-{
-  return segment_manager.read(
-    record_base.add_offset(header.mdlength),
-    header.dlength
-  ).safe_then([=, &header](auto bptr) {
-    bufferlist bl;
-    bl.append(bptr);
-    return bl.crc32c(-1) == header.data_crc;
-  });
 }
 
 Journal::write_record_ret Journal::write_record(
@@ -215,49 +186,6 @@ Journal::roll_journal_segment()
     );
 }
 
-Journal::read_segment_header_ret
-Journal::read_segment_header(segment_id_t segment)
-{
-  return segment_manager.read(
-    paddr_t{segment, 0},
-    segment_manager.get_block_size()
-  ).handle_error(
-    read_segment_header_ertr::pass_further{},
-    crimson::ct_error::assert_all{
-      "Invalid error in Journal::read_segment_header"
-    }
-  ).safe_then([=](bufferptr bptr) -> read_segment_header_ret {
-    logger().debug("segment {} bptr size {}", segment, bptr.length());
-
-    segment_header_t header;
-    bufferlist bl;
-    bl.push_back(bptr);
-
-    logger().debug(
-      "Journal::read_segment_header: segment {} block crc {}",
-      segment,
-      bl.begin().crc32c(segment_manager.get_block_size(), 0));
-
-    auto bp = bl.cbegin();
-    try {
-      decode(header, bp);
-    } catch (ceph::buffer::error &e) {
-      logger().debug(
-	"Journal::read_segment_header: segment {} unable to decode "
-	"header, skipping",
-	segment);
-      return crimson::ct_error::enodata::make();
-    }
-    logger().debug(
-      "Journal::read_segment_header: segment {} header {}",
-      segment,
-      header);
-    return read_segment_header_ret(
-      read_segment_header_ertr::ready_future_marker{},
-      header);
-  });
-}
-
 Journal::open_for_write_ret Journal::open_for_write()
 {
   return roll_journal_segment().safe_then([this](auto seq) {
@@ -272,187 +200,79 @@ Journal::open_for_write_ret Journal::open_for_write()
   });
 }
 
-Journal::find_replay_segments_fut Journal::find_replay_segments()
+Journal::prep_replay_segments_fut
+Journal::prep_replay_segments(
+  std::vector<std::pair<segment_id_t, segment_header_t>> segments)
 {
-  return seastar::do_with(
-    std::vector<std::pair<segment_id_t, segment_header_t>>(),
-    [this](auto &&segments) mutable {
-      return crimson::do_for_each(
-	boost::make_counting_iterator(segment_id_t{0}),
-	boost::make_counting_iterator(segment_manager.get_num_segments()),
-	[this, &segments](auto i) {
-	  return read_segment_header(i
-	  ).safe_then([this, &segments, i](auto header) mutable {
-	    if (generate_nonce(
-		  header.journal_segment_seq,
-		  segment_manager.get_meta()) != header.segment_nonce) {
-	      logger().debug(
-		"find_replay_segments: nonce mismatch segment {} header {}",
-		i,
-		header);
-	      assert(0 == "impossible");
-	      return find_replay_segments_ertr::now();
-	    }
-
-	    segments.emplace_back(i, std::move(header));
-	    return find_replay_segments_ertr::now();
-	  }).handle_error(
-	    crimson::ct_error::enoent::handle([i](auto) {
-	      logger().debug(
-		"find_replay_segments: segment {} not available for read",
-		i);
-	      return find_replay_segments_ertr::now();
-	    }),
-	    crimson::ct_error::enodata::handle([i](auto) {
-	      logger().debug(
-		"find_replay_segments: segment {} header undecodable",
-		i);
-	      return find_replay_segments_ertr::now();
-	    }),
-	    find_replay_segments_ertr::pass_further{},
-	    crimson::ct_error::assert_all{
-	      "Invalid error in Journal::find_replay_segments"
-            }
-	  );
-	}).safe_then([this, &segments]() mutable -> find_replay_segments_fut {
-	  logger().debug(
-	    "find_replay_segments: have {} segments",
-	    segments.size());
-	  if (segments.empty()) {
-	    return crimson::ct_error::input_output_error::make();
-	  }
-	  std::sort(
-	    segments.begin(),
-	    segments.end(),
-	    [](const auto &lt, const auto &rt) {
-	      return lt.second.journal_segment_seq <
-		rt.second.journal_segment_seq;
-	    });
-
-	  next_journal_segment_seq =
-	    segments.rbegin()->second.journal_segment_seq + 1;
-	  std::for_each(
-	    segments.begin(),
-	    segments.end(),
-	    [this](auto &seg) {
-	      segment_provider->init_mark_segment_closed(
-		seg.first,
-		seg.second.journal_segment_seq);
-	    });
-
-	  auto journal_tail = segments.rbegin()->second.journal_tail;
-	  segment_provider->update_journal_tail_committed(journal_tail);
-	  auto replay_from = journal_tail.offset;
-	  logger().debug(
-	    "Journal::find_replay_segments: journal_tail={}",
-	    journal_tail);
-	  auto from = segments.begin();
-	  if (replay_from != P_ADDR_NULL) {
-	    from = std::find_if(
-	      segments.begin(),
-	      segments.end(),
-	      [&replay_from](const auto &seg) -> bool {
-		return seg.first == replay_from.segment;
-	      });
-	    if (from->second.journal_segment_seq != journal_tail.segment_seq) {
-	      logger().error(
-		"find_replay_segments: journal_tail {} does not match {}",
-		journal_tail,
-		from->second);
-	      assert(0 == "invalid");
-	    }
-	  } else {
-	    replay_from = paddr_t{
-	      from->first,
-	      (segment_off_t)segment_manager.get_block_size()};
-	  }
-	  auto ret = replay_segments_t(segments.end() - from);
-	  std::transform(
-	    from, segments.end(), ret.begin(),
-	    [this](const auto &p) {
-	      auto ret = journal_seq_t{
-		p.second.journal_segment_seq,
-		paddr_t{
-		  p.first,
-		  (segment_off_t)segment_manager.get_block_size()}};
-	      logger().debug(
-		"Journal::find_replay_segments: replaying from  {}",
-		ret);
-	      return std::make_pair(ret, p.second);
-	    });
-	  ret[0].first.offset = replay_from;
-	  return find_replay_segments_fut(
-	    find_replay_segments_ertr::ready_future_marker{},
-	    std::move(ret));
-	});
-    });
-}
-
-Journal::read_validate_record_metadata_ret Journal::read_validate_record_metadata(
-  paddr_t start,
-  segment_nonce_t nonce)
-{
-  auto block_size = segment_manager.get_block_size();
-  if (start.offset + block_size > (int64_t)segment_manager.get_segment_size()) {
-    return read_validate_record_metadata_ret(
-      read_validate_record_metadata_ertr::ready_future_marker{},
-      std::nullopt);
+  logger().debug(
+    "prep_replay_segments: have {} segments",
+    segments.size());
+  if (segments.empty()) {
+    return crimson::ct_error::input_output_error::make();
   }
-  return segment_manager.read(start, block_size
-  ).safe_then(
-    [=](bufferptr bptr) mutable
-    -> read_validate_record_metadata_ret {
-      logger().debug("read_validate_record_metadata: reading {}", start);
-      auto block_size = segment_manager.get_block_size();
-      bufferlist bl;
-      bl.append(bptr);
-      auto bp = bl.cbegin();
-      record_header_t header;
-      try {
-	decode(header, bp);
-      } catch (ceph::buffer::error &e) {
-	return read_validate_record_metadata_ret(
-	  read_validate_record_metadata_ertr::ready_future_marker{},
-	  std::nullopt);
-      }
-      if (header.segment_nonce != nonce) {
-	return read_validate_record_metadata_ret(
-	  read_validate_record_metadata_ertr::ready_future_marker{},
-	  std::nullopt);
-      }
-      if (header.mdlength > (extent_len_t)block_size) {
-	if (start.offset + header.mdlength >
-	    (int64_t)segment_manager.get_segment_size()) {
-	  return crimson::ct_error::input_output_error::make();
-	}
-	return segment_manager.read(
-	  {start.segment, start.offset + (segment_off_t)block_size},
-	  header.mdlength - block_size).safe_then(
-	    [header=std::move(header), bl=std::move(bl)](
-	      auto &&bptail) mutable {
-	      bl.push_back(bptail);
-	      return read_validate_record_metadata_ret(
-		read_validate_record_metadata_ertr::ready_future_marker{},
-		std::make_pair(std::move(header), std::move(bl)));
-	    });
-      } else {
-	return read_validate_record_metadata_ret(
-	  read_validate_record_metadata_ertr::ready_future_marker{},
-	  std::make_pair(std::move(header), std::move(bl))
-	);
-      }
-    }).safe_then([=](auto p) {
-      if (p && validate_metadata(p->second)) {
-	return read_validate_record_metadata_ret(
-	  read_validate_record_metadata_ertr::ready_future_marker{},
-	  std::move(*p)
-	);
-      } else {
-	return read_validate_record_metadata_ret(
-	  read_validate_record_metadata_ertr::ready_future_marker{},
-	  std::nullopt);
-      }
+  std::sort(
+    segments.begin(),
+    segments.end(),
+    [](const auto &lt, const auto &rt) {
+      return lt.second.journal_segment_seq <
+	rt.second.journal_segment_seq;
     });
+
+  next_journal_segment_seq =
+    segments.rbegin()->second.journal_segment_seq + 1;
+  std::for_each(
+    segments.begin(),
+    segments.end(),
+    [this](auto &seg) {
+      segment_provider->init_mark_segment_closed(
+	seg.first,
+	seg.second.journal_segment_seq);
+    });
+
+  auto journal_tail = segments.rbegin()->second.journal_tail;
+  segment_provider->update_journal_tail_committed(journal_tail);
+  auto replay_from = journal_tail.offset;
+  logger().debug(
+    "Journal::prep_replay_segments: journal_tail={}",
+    journal_tail);
+  auto from = segments.begin();
+  if (replay_from != P_ADDR_NULL) {
+    from = std::find_if(
+      segments.begin(),
+      segments.end(),
+      [&replay_from](const auto &seg) -> bool {
+	return seg.first == replay_from.segment;
+      });
+    if (from->second.journal_segment_seq != journal_tail.segment_seq) {
+      logger().error(
+	"prep_replay_segments: journal_tail {} does not match {}",
+	journal_tail,
+	from->second);
+      assert(0 == "invalid");
+    }
+  } else {
+    replay_from = paddr_t{
+      from->first,
+      (segment_off_t)segment_manager.get_block_size()};
+  }
+  auto ret = replay_segments_t(segments.end() - from);
+  std::transform(
+    from, segments.end(), ret.begin(),
+    [this](const auto &p) {
+      auto ret = journal_seq_t{
+	p.second.journal_segment_seq,
+	paddr_t{
+	  p.first,
+	  (segment_off_t)segment_manager.get_block_size()}};
+      logger().debug(
+	"Journal::prep_replay_segments: replaying from  {}",
+	ret);
+      return std::make_pair(ret, p.second);
+    });
+  ret[0].first.offset = replay_from;
+  return prep_replay_segments_fut(
+    prep_replay_segments_ertr::ready_future_marker{},
+    std::move(ret));
 }
 
 std::optional<std::vector<delta_info_t>> Journal::try_decode_deltas(
@@ -475,25 +295,6 @@ std::optional<std::vector<delta_info_t>> Journal::try_decode_deltas(
   return deltas;
 }
 
-std::optional<std::vector<extent_info_t>> Journal::try_decode_extent_infos(
-  record_header_t header,
-  const bufferlist &bl)
-{
-  auto bliter = bl.cbegin();
-  bliter += ceph::encoded_sizeof_bounded<record_header_t>();
-  bliter += sizeof(checksum_t) /* crc */;
-  logger().debug("{}: decoding {} extents", __func__, header.extents);
-  std::vector<extent_info_t> extent_infos(header.extents);
-  for (auto &&i : extent_infos) {
-    try {
-      decode(i, bliter);
-    } catch (ceph::buffer::error &e) {
-      return std::nullopt;
-    }
-  }
-  return extent_infos;
-}
-
 Journal::replay_ertr::future<>
 Journal::replay_segment(
   journal_seq_t seq,
@@ -503,7 +304,7 @@ Journal::replay_segment(
   logger().debug("replay_segment: starting at {}", seq);
   return seastar::do_with(
     scan_valid_records_cursor(seq.offset),
-    found_record_handler_t(
+    Scanner::found_record_handler_t(
       [=, &handler](paddr_t base,
 		    const record_header_t &header,
 		    const bufferlist &mdbuf) {
@@ -548,20 +349,29 @@ Journal::replay_segment(
 	  });
       }),
     [=](auto &cursor, auto &dhandler) {
-      return scan_valid_records(
+      return scanner.scan_valid_records(
 	cursor,
 	header.segment_nonce,
 	std::numeric_limits<size_t>::max(),
-	dhandler).safe_then([](auto){});
+	dhandler).safe_then([](auto){}
+      ).handle_error(
+	replay_ertr::pass_further{},
+	crimson::ct_error::assert_all{
+	  "shouldn't meet with any other error other replay_ertr"
+	}
+      );;
     });
 }
 
-Journal::replay_ret Journal::replay(delta_handler_t &&delta_handler)
+Journal::replay_ret Journal::replay(
+  std::vector<std::pair<segment_id_t, segment_header_t>>&& segment_headers,
+  delta_handler_t &&delta_handler)
 {
   return seastar::do_with(
     std::move(delta_handler), replay_segments_t(),
-    [this](auto &handler, auto &segments) mutable -> replay_ret {
-      return find_replay_segments().safe_then(
+    [this, segment_headers=std::move(segment_headers)]
+    (auto &handler, auto &segments) mutable -> replay_ret {
+      return prep_replay_segments(std::move(segment_headers)).safe_then(
         [this, &handler, &segments](auto replay_segs) mutable {
           logger().debug("replay: found {} segments", replay_segs.size());
           segments = std::move(replay_segs);
@@ -571,166 +381,5 @@ Journal::replay_ret Journal::replay(delta_handler_t &&delta_handler)
         });
     });
 }
-
-Journal::scan_extents_ret Journal::scan_extents(
-  scan_extents_cursor &cursor,
-  extent_len_t bytes_to_read)
-{
-  auto ret = std::make_unique<scan_extents_ret_bare>();
-  auto* extents = ret.get();
-  return read_segment_header(cursor.get_offset().segment
-  ).handle_error(
-    scan_extents_ertr::pass_further{},
-    crimson::ct_error::assert_all{
-      "Invalid error in Journal::scan_extents"
-    }
-  ).safe_then([bytes_to_read, extents, &cursor, this](auto segment_header) {
-    auto segment_nonce = segment_header.segment_nonce;
-    return seastar::do_with(
-      found_record_handler_t(
-	[extents, this](
-	  paddr_t base,
-	  const record_header_t &header,
-	  const bufferlist &mdbuf) mutable {
-
-	  auto infos = try_decode_extent_infos(
-	    header,
-	    mdbuf);
-	  if (!infos) {
-	    // This should be impossible, we did check the crc on the mdbuf
-	    logger().error(
-	      "Journal::scan_extents unable to decode extents for record {}",
-	      base);
-	    assert(infos);
-	  }
-
-	  paddr_t extent_offset = base.add_offset(header.mdlength);
-	  for (const auto &i : *infos) {
-	    extents->emplace_back(extent_offset, i);
-	    extent_offset.offset += i.len;
-	  }
-	  return scan_extents_ertr::now();
-	}),
-      [=, &cursor](auto &dhandler) {
-	return scan_valid_records(
-	  cursor,
-	  segment_nonce,
-	  bytes_to_read,
-	  dhandler).discard_result();
-      });
-  }).safe_then([ret=std::move(ret)] {
-    return std::move(*ret);
-  });
-}
-
-Journal::scan_valid_records_ret Journal::scan_valid_records(
-    scan_valid_records_cursor &cursor,
-    segment_nonce_t nonce,
-    size_t budget,
-    found_record_handler_t &handler)
-{
-  if (cursor.offset.offset == 0) {
-    cursor.offset.offset = segment_manager.get_block_size();
-  }
-  auto retref = std::make_unique<size_t>(0);
-  auto budget_used = *retref;
-  return crimson::repeat(
-    [=, &cursor, &budget_used, &handler]() mutable
-    -> scan_valid_records_ertr::future<seastar::stop_iteration> {
-      return [=, &handler, &cursor, &budget_used] {
-	if (!cursor.last_valid_header_found) {
-	  return read_validate_record_metadata(cursor.offset, nonce
-	  ).safe_then([=, &cursor](auto md) {
-	    logger().debug(
-	      "Journal::scan_valid_records: read complete {}",
-	      cursor.offset);
-	    if (!md) {
-	      logger().debug(
-		"Journal::scan_valid_records: found invalid header at {}, presumably at end",
-		cursor.offset);
-	      cursor.last_valid_header_found = true;
-	      return scan_valid_records_ertr::now();
-	    } else {
-	      logger().debug(
-		"Journal::scan_valid_records: valid record read at {}",
-		cursor.offset);
-	      cursor.last_committed = paddr_t{
-		cursor.offset.segment,
-		md->first.committed_to};
-	      cursor.pending_records.emplace_back(
-		cursor.offset,
-		md->first,
-		md->second);
-	      cursor.offset.offset +=
-		md->first.dlength + md->first.mdlength;
-	      return scan_valid_records_ertr::now();
-	    }
-	  }).safe_then([=, &cursor, &budget_used, &handler] {
-	    return crimson::repeat(
-	      [=, &budget_used, &cursor, &handler] {
-		logger().debug(
-		  "Journal::scan_valid_records: valid record read, processing queue");
-		if (cursor.pending_records.empty()) {
-		  /* This is only possible if the segment is empty.
-		   * A record's last_commited must be prior to its own
-		   * location since it itself cannot yet have been committed
-		   * at its own time of submission.  Thus, the most recently
-		   * read record must always fall after cursor.last_committed */
-		  return scan_valid_records_ertr::make_ready_future<
-		    seastar::stop_iteration>(seastar::stop_iteration::yes);
-		}
-		auto &next = cursor.pending_records.front();
-		if (next.offset > cursor.last_committed) {
-		  return scan_valid_records_ertr::make_ready_future<
-		    seastar::stop_iteration>(seastar::stop_iteration::yes);
-		}
-		budget_used +=
-		  next.header.dlength + next.header.mdlength;
-		return handler(
-		  next.offset,
-		  next.header,
-		  next.mdbuffer
-		).safe_then([&cursor] {
-		  cursor.pending_records.pop_front();
-		  return scan_valid_records_ertr::make_ready_future<
-		    seastar::stop_iteration>(seastar::stop_iteration::no);
-		});
-	      });
-	  });
-	} else {
-	  assert(!cursor.pending_records.empty());
-	  auto &next = cursor.pending_records.front();
-	  return read_validate_data(next.offset, next.header
-	  ).safe_then([=, &budget_used, &next, &cursor, &handler](auto valid) {
-	    if (!valid) {
-	      cursor.pending_records.clear();
-	      return scan_valid_records_ertr::now();
-	    }
-	    budget_used +=
-	      next.header.dlength + next.header.mdlength;
-	    return handler(
-	      next.offset,
-	      next.header,
-	      next.mdbuffer
-	    ).safe_then([&cursor] {
-	      cursor.pending_records.pop_front();
-	      return scan_valid_records_ertr::now();
-	    });
-	  });
-	}
-      }().safe_then([=, &budget_used, &cursor] {
-	if (cursor.is_complete() || budget_used >= budget) {
-	  return seastar::stop_iteration::yes;
-	} else {
-	  return seastar::stop_iteration::no;
-	}
-      });
-    }).safe_then([retref=std::move(retref)]() mutable -> scan_valid_records_ret {
-      return scan_valid_records_ret(
-	scan_valid_records_ertr::ready_future_marker{},
-	std::move(*retref));
-    });
-}
-
 
 }
