+++ src/crimson/os/seastore/journal.h
@@ -14,6 +14,7 @@
 #include "include/denc.h"
 
 #include "crimson/common/log.h"
+#include "crimson/os/seastore/scanner.h"
 #include "crimson/os/seastore/segment_manager.h"
 #include "crimson/os/seastore/ordering_handle.h"
 #include "crimson/os/seastore/seastore_types.h"
@@ -29,7 +30,7 @@ class SegmentedAllocator;
  */
 class Journal {
 public:
-  Journal(SegmentManager &segment_manager);
+  Journal(SegmentManager &segment_manager, Scanner& scanner);
 
   /**
    * Sets the SegmentProvider.
@@ -136,26 +137,9 @@ public:
     replay_ret(journal_seq_t seq,
 	       paddr_t record_block_base,
 	       const delta_info_t&)>;
-  replay_ret replay(delta_handler_t &&delta_handler);
-
-  /**
-   * scan_extents
-   *
-   * Scans records beginning at addr until the first record boundary after
-   * addr + bytes_to_read.
-   *
-   * Returns list<extent, extent_info>
-   * cursor.is_complete() will be true when no further extents exist in segment.
-   */
-  class scan_valid_records_cursor;
-  using scan_extents_cursor = scan_valid_records_cursor;
-  using scan_extents_ertr = SegmentManager::read_ertr;
-  using scan_extents_ret_bare = std::list<std::pair<paddr_t, extent_info_t>>;
-  using scan_extents_ret = scan_extents_ertr::future<scan_extents_ret_bare>;
-  scan_extents_ret scan_extents(
-    scan_extents_cursor &cursor,
-    extent_len_t bytes_to_read
-  );
+  replay_ret replay(
+    std::vector<std::pair<segment_id_t, segment_header_t>>&& segment_headers,
+    delta_handler_t &&delta_handler);
 
   void set_write_pipeline(WritePipeline *_write_pipeline) {
     write_pipeline = _write_pipeline;
@@ -172,6 +156,7 @@ private:
   segment_off_t written_to = 0;
   segment_off_t committed_to = 0;
 
+  Scanner& scanner;
   WritePipeline *write_pipeline = nullptr;
 
   void reset_soft_state() {
@@ -186,18 +171,6 @@ private:
     crimson::ct_error::input_output_error>;
   initialize_segment_ertr::future<segment_seq_t> initialize_segment(
     Segment &segment);
-
-  /// validate embedded metadata checksum
-  static bool validate_metadata(const bufferlist &bl);
-
-  /// read and validate data
-  using read_validate_data_ertr = SegmentManager::read_ertr;
-  using read_validate_data_ret = read_validate_data_ertr::future<bool>;
-  read_validate_data_ret read_validate_data(
-    paddr_t record_base,
-    const record_header_t &header  ///< caller must ensure lifetime through
-                                   ///  future resolution
-  );
 
 
   /// do record write
@@ -217,95 +190,23 @@ private:
   /// returns true iff current segment has insufficient space
   bool needs_roll(segment_off_t length) const;
 
-  using read_segment_header_ertr = crimson::errorator<
-    crimson::ct_error::enoent,
-    crimson::ct_error::enodata,
-    crimson::ct_error::input_output_error
-    >;
-  using read_segment_header_ret = read_segment_header_ertr::future<
-    segment_header_t>;
-  read_segment_header_ret read_segment_header(segment_id_t segment);
-
   /// return ordered vector of segments to replay
   using replay_segments_t = std::vector<
     std::pair<journal_seq_t, segment_header_t>>;
-  using find_replay_segments_ertr = crimson::errorator<
+  using prep_replay_segments_ertr = crimson::errorator<
     crimson::ct_error::input_output_error
     >;
-  using find_replay_segments_fut = find_replay_segments_ertr::future<
+  using prep_replay_segments_fut = prep_replay_segments_ertr::future<
     replay_segments_t>;
-  find_replay_segments_fut find_replay_segments();
+  prep_replay_segments_fut prep_replay_segments(
+    std::vector<std::pair<segment_id_t, segment_header_t>> segments);
 
   /// attempts to decode deltas from bl, return nullopt if unsuccessful
   std::optional<std::vector<delta_info_t>> try_decode_deltas(
     record_header_t header,
     const bufferlist &bl);
 
-  /// attempts to decode extent infos from bl, return nullopt if unsuccessful
-  std::optional<std::vector<extent_info_t>> try_decode_extent_infos(
-    record_header_t header,
-    const bufferlist &bl);
-
-  /// read record metadata for record starting at start
-  using read_validate_record_metadata_ertr = replay_ertr;
-  using read_validate_record_metadata_ret =
-    read_validate_record_metadata_ertr::future<
-      std::optional<std::pair<record_header_t, bufferlist>>
-    >;
-  read_validate_record_metadata_ret read_validate_record_metadata(
-    paddr_t start,
-    segment_nonce_t nonce);
-
-public:
-  /// scan segment for end incrementally
-  struct scan_valid_records_cursor {
-    bool last_valid_header_found = false;
-    paddr_t offset;
-    paddr_t last_committed;
-
-    struct found_record_t {
-      paddr_t offset;
-      record_header_t header;
-      bufferlist mdbuffer;
-
-      found_record_t(
-	paddr_t offset,
-	const record_header_t &header,
-	const bufferlist &mdbuffer)
-	: offset(offset), header(header), mdbuffer(mdbuffer) {}
-    };
-    std::deque<found_record_t> pending_records;
-
-    bool is_complete() const {
-      return last_valid_header_found && pending_records.empty();
-    }
-
-    paddr_t get_offset() const {
-      return offset;
-    }
-
-    scan_valid_records_cursor(
-      paddr_t offset)
-      : offset(offset) {}
-  };
 private:
-
-  using scan_valid_records_ertr = SegmentManager::read_ertr;
-  using scan_valid_records_ret = scan_valid_records_ertr::future<
-    size_t>;
-  using found_record_handler_t = std::function<
-    scan_valid_records_ertr::future<>(
-      paddr_t record_block_base,
-      // callee may assume header and bl will remain valid until
-      // returned future resolves
-      const record_header_t &header,
-      const bufferlist &bl)>;
-  scan_valid_records_ret scan_valid_records(
-    scan_valid_records_cursor &cursor, ///< [in, out] cursor, updated during call
-    segment_nonce_t nonce,             ///< [in] nonce for segment
-    size_t budget,                     ///< [in] max budget to use
-    found_record_handler_t &handler    ///< [in] handler for records
-  ); ///< @return used budget
 
   /// replays records starting at start through end of segment
   replay_ertr::future<>
